Did a major refactoring on the drawing system. 	*DrawingInfo are all Interfaces which are implemented by *DrawingInfoImpl these Interfaces are used for different parts in the drawing process. 	Therefore LifelineSpanningTickSpanningOccurrence had to be changed to take advantage of this new classes. 	Drawing code is more clean and does not contain any index access with the lifeline.getIndex(). 	Added a padding between ticks so that elements which need the full tick height don't stick together
Combined fragments have now a horizontal border padding which is needed for nested fragments.
(the usable lifeline width of the starting and ending lifelines shrink in the fragment)

Next step is adding a vertical border padding, doing another refactoring because the planned Solution with the vertical tick padding won't work.

FYI:LifelineOccurrence wasn't changed because the Lifeline did and does the math to calculate the position and size.