Add an "Export as" panel to the menu of UMLet standalone (Swing edition) for quick export of the current diagram
The changes in this commit extend the menu bar with an "Export as" panel that
comprises a combobox with the export formats supported by UMLet, i.e., the same
file formats available at the File menu in its "Export as..." entry. Furthermore,
the new panel also contains an "Export" button. Next to clicking it, the button
may also be activated by hitting Ctrl+E.

The "Export" button behaves slightly different depending on the current diagram
tab's "export status":
  1. If the diagram has not been exported before, a file chooser for the export
     destination is openened. That is, the button behaves exactly like the
     file format entries under the "Export as..." submenu.
  2. If the diagram has been exported the button's behavior depends on the
     following circumstances:
       2.1 The user did not change the export format extension in the combo
           box and did not export the diagram leveraging the "Export as..."
           submenu: The diagram gets directly exported to the destination it
           was exported the last time. Note, that there is no warning in case
           the previously exported file exists, i.e., it may be overwritten
           without any warning.
       2.2 The user changed the export format extension or exported the diagram
           via the menu, whereby she changed the name of the export file: The
           button behaves like the file format entries under the "Export as..."
           submenu (same behavior as described above in 1.).

When the user opens a new tab, either by clicking an "File" --> "New" or by
opening a bunch of files via "File" --> "Open...", the format in the combo box
will be set to "PDF", which is the default export format. In contrast, if a
user switches from one diagram tab to another and the diagram has been exported
before, the combo box will be set to the file format of the last export file.