UVC-031: Modified calculation of lineGap
If the bbox includes all available characters (and therefore their bounds are larger than the difference of the specified ascender and descender), then use it to calculate the lineGap.
If this is not the case, approximate the lineGap as 2/3 of the distance between the fonts baseline and the minimum y value.