Umlet unix wrapper: Allow to symlink the umtlet.sh
The default umlet.sh script is using a trick to resolve the default umlet path
but this trick is not working well when using symbolic links. In this patch
realink is used to resolve the path of the shell script instead of $0 directly

With this change in in place I was able to test that creating a symbolic link
from a bin directory to the umlet.sh wrapper now works nicely but the old
behaviour: adding umlet.sh into the path is also still functional.

Tested on Ubuntu 22.04 with UMLet version 15.1 2023-03-12